hotfix: fetch retry fix (#8615)
